added some thingies

+++ migrations/20240702190330_init.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS "Accounts" (
 	"banned"	    INTEGER NOT NULL DEFAULT 0,
 	"mine_slots"	INTEGER NOT NULL DEFAULT 0, --number of mine slots
 	"previous_claim"	INTEGER NOT NULL DEFAULT 0, --unix timestamp
-	"awaiting_claim"	INTEGER NOT NULL DEFAULT 0, --number of mathcoins that have not been claimed
+	"awaiting_claim"	REAL NOT NULL DEFAULT 0.0, --number of mathcoins that have not been claimed
 	"username"		TEXT NOT NULL COLLATE NOCASE UNIQUE,
 	"user_bio"		TEXT, --optional
 	"pronouns"		TEXT, --optional
